Extract U32 texture data upload logic and restore pixelStorei parameters
