+++ src/utils.ts
@@ -1676,3 +1676,44 @@ export class GunzipReader {
     return result;
   }
 }
+
+export function uploadU32DataTextureRows(
+  renderer: THREE.WebGLRenderer,
+  texture: THREE.Texture,
+  width: number,
+  rows: number,
+  data: Uint32Array,
+) {
+  const gl = renderer.getContext() as WebGL2RenderingContext;
+
+  const props = renderer.properties.get(texture) as {
+    __webglTexture: WebGLTexture;
+  };
+  const glTexture = props?.__webglTexture;
+  if (!glTexture) {
+    throw new Error("texture not found");
+  }
+  // Note: instead of saving and restoring the pixelStorei parameters
+  //       renderer.state.pixelStorei can be used with Three.js >= r184
+  const currentFlipY = gl.getParameter(gl.UNPACK_FLIP_Y_WEBGL);
+  const currentPremultiply = gl.getParameter(gl.UNPACK_PREMULTIPLY_ALPHA_WEBGL);
+  renderer.state.activeTexture(gl.TEXTURE0);
+  renderer.state.bindTexture(gl.TEXTURE_2D, glTexture);
+  gl.bindBuffer(gl.PIXEL_UNPACK_BUFFER, null);
+  gl.pixelStorei(gl.UNPACK_FLIP_Y_WEBGL, false);
+  gl.pixelStorei(gl.UNPACK_PREMULTIPLY_ALPHA_WEBGL, false);
+  gl.texSubImage2D(
+    gl.TEXTURE_2D,
+    0,
+    0,
+    0,
+    width,
+    rows,
+    gl.RGBA_INTEGER,
+    gl.UNSIGNED_INT,
+    data,
+  );
+  renderer.state.unbindTexture();
+  gl.pixelStorei(gl.UNPACK_FLIP_Y_WEBGL, currentFlipY);
+  gl.pixelStorei(gl.UNPACK_PREMULTIPLY_ALPHA_WEBGL, currentPremultiply);
+}
